Reorganized hooks
