+++ apps/frontend-react/src/features/developerPanel/TransmissionDetails.tsx
@@ -18,7 +18,7 @@ import { IconAdjustments, IconTrash } from "@tabler/icons-react";
 import { useState } from "react";
 import AudioPlayer, { RHAP_UI } from "react-h5-audio-player";
 import "react-h5-audio-player/lib/styles.css";
-import useQueryTransmission from "../../queries/useQueryTransmission";
+import useQueryTransmission from "../../hooks/useQueryTransmission";
 import TransmissionQueries from "../../queries/TransmissionQueries";
 
 //
