+++ apps/frontend-react/src/hooks/useStreamSocket.ts
